# api-spec.yaml
openapi: 3.0.3
info:
  title: AgentHub Registry API
  description: |
    The universal package registry for AI-native agents, tools, chains, and prompts.

    ## Authentication

    This API uses GitHub OAuth 2.0 for authentication. Most endpoints require a valid JWT token.

    ## Rate Limiting

    API requests are rate limited to 1000 requests per minute per IP address.

    ## Package Types

    The registry supports the following package types:
    - `agent`: AI agents and assistants
    - `tool`: Utilities and functions for agents
    - `chain`: Workflow chains and pipelines
    - `prompt`: Prompt templates and examples
    - `dataset`: Training and reference datasets

    ## Versioning

    All packages follow semantic versioning (semver) conventions.
  version: 1.0.0
  contact:
    name: AgentHub Team
    url: https://github.com/agenthubcli/registry
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://registry.agenthubcli.com
    description: Production server
  - url: http://localhost:8000
    description: Development server
paths:
...
  /api/v1/packages/publish:
    post:
      summary: Publish package
      description: Upload and publish a new package or version
      tags: [Packages]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PackageUpload"
      responses:
        "200":
          description: Package published successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishSuccess"
        "400":
          description: Invalid package data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: User not authorized to publish
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Package version already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Package too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/packages/{package_name}:
    get:
      summary: Get package details
      description: Get detailed information about a package including all versions
      tags: [Packages]
      parameters:
        - $ref: "#/components/parameters/PackageName"
      responses:
        "200":
          description: Package details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageDetails"
        "404":
          description: Package not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      summary: Delete package
      description: Delete a package and all its versions (admin or owner only)
      tags: [Packages]
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PackageName"
      responses:
        "200":
          description: Package deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not authorized to delete package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Package not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/packages/{package_name}/versions:
    get:
      summary: Get package versions
      description: Get all versions of a package
      tags: [Packages]
      parameters:
        - $ref: "#/components/parameters/PackageName"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Package versions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageVersions"
        "404":
          description: Package not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/packages/{package_name}/{version}:
    get:
      summary: Get specific version
      description: Get details about a specific package version
      tags: [Packages]
      parameters:
        - $ref: "#/components/parameters/PackageName"
        - $ref: "#/components/parameters/Version"
      responses:
        "200":
          description: Package version details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageVersionDetails"
        "404":
          description: Package or version not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/packages/{package_name}/{version}/download:
    get:
      summary: Download package
      description: Download a specific package version
      tags: [Packages]
      parameters:
        - $ref: "#/components/parameters/PackageName"
        - $ref: "#/components/parameters/Version"
      responses:
        "200":
          description: Package file
          headers:
            Content-Disposition:
              description: Attachment filename
              schema:
                type: string
            Content-Length:
              description: File size in bytes
              schema:
                type: integer
            X-Package-Name:
              description: Package name
              schema:
                type: string
            X-Package-Version:
              description: Package version
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Package, version, or file not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/packages/{package_name}/stats:
    get:
      summary: Get package statistics
      description: Get download and usage statistics for a package
      tags: [Packages]
      parameters:
        - $ref: "#/components/parameters/PackageName"
      responses:
        "200":
          description: Package statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageStats"
        "404":
          description: Package not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # Search Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained via GitHub OAuth
  parameters:
    PackageName:
      name: package_name
      in: path
      required: true
      description: Package name
      schema:
        type: string
        pattern: '^[a-z0-9]([a-z0-9\-])*[a-z0-9]$'
        minLength: 1
        maxLength: 214
      example: "my-awesome-tool"
    Version:
      name: version
      in: path
      required: true
      description: Package version (semver)
      schema:
        type: string
        pattern: '^[0-9]+\.[0-9]+\.[0-9]+([+-][a-zA-Z0-9\-\.]+)*$'
      example: "1.2.3"
...
    Limit:
      name: limit
      in: query
      description: Maximum number of results
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      description: Number of results to skip
      schema:
        type: integer
        minimum: 0
        default: 0
  schemas:
...
    PackageType:
      type: string
      enum: [agent, tool, chain, prompt, dataset]
      description: Type of package

    # Common Responses
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Error message
        error_code:
          type: string
          description: Machine-readable error code
      example:
        detail: "Package not found"
        error_code: "PACKAGE_NOT_FOUND"
    Message:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Success message
      example:
        message: "Operation completed successfully"

    # API Info
...
    UserProfile:
      type: object
      required: [id, github_username]
      properties:
        id:
          type: integer
          example: 123
        github_username:
          type: string
          example: "johndoe"
        display_name:
          type: string
          example: "John Doe"
        github_avatar_url:
          type: string
          format: uri
          example: "https://github.com/avatars/johndoe"
        bio:
          type: string
          example: "AI researcher and developer"
        website:
          type: string
          format: uri
          example: "https://johndoe.dev"
        location:
          type: string
          example: "San Francisco, CA"
        company:
          type: string
          example: "AI Company Inc."
        total_packages:
          type: integer
          example: 5
        total_downloads:
          type: integer
          example: 1234
        created_at:
          type: string
          format: date-time

    # Package Upload
    PackageUpload:
      type: object
      required: [file, package_type]
      properties:
        file:
          type: string
          format: binary
          description: Package file (tar.gz, zip, etc.)
        package_type:
          $ref: "#/components/schemas/PackageType"
    PublishSuccess:
      type: object
      required: [message, package, version]
      properties:
        message:
          type: string
          example: "Package published successfully"
        package:
          $ref: "#/components/schemas/Package"
        version:
          $ref: "#/components/schemas/PackageVersion"

    # Package
    Package:
      type: object
      required: [id, name, package_type, owner, created_at]
      properties:
        id:
          type: integer
          example: 456
        name:
          type: string
          example: "awesome-tool"
        description:
          type: string
          example: "An awesome tool for AI development"
        package_type:
          $ref: "#/components/schemas/PackageType"
        latest_version:
          type: string
          example: "1.2.3"
        total_downloads:
          type: integer
          example: 5678
        download_count_last_30_days:
          type: integer
          example: 234
        version_count:
          type: integer
          example: 8
        homepage:
          type: string
          format: uri
          example: "https://github.com/user/awesome-tool"
        repository:
          type: string
          format: uri
          example: "https://github.com/user/awesome-tool"
        documentation:
          type: string
          format: uri
          example: "https://awesome-tool.readthedocs.io"
        keywords:
          type: array
          items:
            type: string
          example: ["ai", "tool", "automation"]
        owner:
          $ref: "#/components/schemas/UserProfile"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PackageVersion:
      type: object
      required: [id, version, download_count, file_size, created_at]
      properties:
        id:
          type: integer
          example: 789
        version:
          type: string
          example: "1.2.3"
        description:
          type: string
          example: "Added new features and bug fixes"
        download_count:
          type: integer
          example: 123
        file_size:
          type: integer
          description: File size in bytes
          example: 2048576
        file_hash_sha256:
          type: string
          example: "abc123..."
        is_prerelease:
          type: boolean
          example: false
        runtime:
          type: string
          example: "python"
        python_version:
          type: string
          example: ">=3.8"
        download_url:
          type: string
          format: uri
          description: Download URL for this version
        manifest:
          type: object
          description: Package manifest data
        published_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    PackageDetails:
      type: object
      required: [package, owner, versions, latest_version]
      properties:
        package:
          $ref: "#/components/schemas/Package"
        owner:
          $ref: "#/components/schemas/UserProfile"
        versions:
          type: array
          items:
            $ref: "#/components/schemas/PackageVersion"
        latest_version:
          $ref: "#/components/schemas/PackageVersion"
    PackageVersions:
      type: object
      required: [package_name, versions, total]
      properties:
        package_name:
          type: string
          example: "awesome-tool"
        versions:
          type: array
          items:
            $ref: "#/components/schemas/PackageVersion"
        total:
          type: integer
          example: 8
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0
    PackageVersionDetails:
      type: object
      required: [package, version, owner]
      properties:
        package:
          $ref: "#/components/schemas/Package"
        version:
          $ref: "#/components/schemas/PackageVersion"
        owner:
          $ref: "#/components/schemas/UserProfile"
    PackageStats:
      type: object
      required: [package_name, total_downloads, version_count]
      properties:
        package_name:
          type: string
          example: "awesome-tool"
        total_downloads:
          type: integer
          example: 5678
        downloads_last_30_days:
          type: integer
          example: 234
        downloads_last_7_days:
          type: integer
          example: 56
        version_count:
          type: integer
          example: 8
        latest_version:
          type: string
          example: "1.2.3"

    # Search
...